+++ apps/openmw/mwgui/windowmanagerimp.hpp
@@ -294,6 +294,7 @@ namespace MWGui
     void hudNotification(const std::string& title, const std::string& value = std::string(),
         const std::string& icon = std::string(), const std::string& key = std::string()) override;
     void hudExperienceNotification(float amount, const std::string& reason = std::string()) override;
+    void hudDamageNumber(float damage) override;
     void staticMessageBox(const std::string& message) override;
     void removeStaticMessageBox() override;
     void interactiveMessageBox(const std::string& message,